+++ assignments/06-heart/submissions/pedro-pignata/src/HeartRates.java
@@ -71,30 +71,3 @@ class HeartRates {
         return targetMin + " bpm - " + targetMax + " bpm";
     }
 }
-
-public class Main {
-    public static void main(String[] args) {
-        Scanner scanner = new Scanner(System.in);
-
-        System.out.print("Digite o seu primeiro nome: ");
-        String firstName = scanner.next();
-
-        System.out.print("Digite o seu sobrenome: ");
-        String lastName = scanner.next();
-
-        System.out.print("Digite sua data de nascimento (dia, mês e ano separados por espaço): ");
-        int dayOfBirth = scanner.nextInt();
-        int monthOfBirth = scanner.nextInt();
-        int yearOfBirth = scanner.nextInt();
-
-        HeartRates person = new HeartRates(firstName, lastName, dayOfBirth, monthOfBirth, yearOfBirth);
-
-        System.out.println("\nNome: " + person.getFirstName() + " " + person.getLastName());
-        System.out.println("Data de nascimento: " + person.getDayOfBirth() + "/" + person.getMonthOfBirth() + "/" + person.getYearOfBirth());
-        System.out.println("Idade: " + person.calculateAge(2025) + " anos");
-        System.out.println("Frequência cardíaca máxima: " + person.calculateMaxHeartRate() + " bpm");
-        System.out.println("Faixa de frequência cardíaca alvo: " + person.calculateTargetHeartRate());
-
-        scanner.close();
-    }
-}
